# openapi.yaml
openapi: 3.0.1
info:
  title: Deus
  description: A plugin that runs local scripts and manages memories.
  version: 'v1'
servers:
  - url: http://localhost:5004
paths:
  /run_script:
    post:
      operationId: run_script
      summary: Run a script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/runScriptRequest'
      responses:
        "200":
          description: Script run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/runScriptResponse'
  /memories:
    post:
      operationId: create_memory
      summary: Create a memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createMemoryRequest'
      responses:
        "201":
          description: Memory created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createMemoryResponse'
        "409":
          description: Memory with this id already exists
    get:
      operationId: list_memories
      summary: List memories
      responses:
        "200":
          description: List of memory ids
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listMemoriesResponse'
  /memories/{memory_id}:
    get:
      operationId: get_memory
      summary: Get a memory
      parameters:
        - in: path
          name: memory_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Memory details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getMemoryResponse'
        "404":
          description: Memory not found
components:
  schemas:
    runScriptRequest:
      type: object
      properties:
        language:
          type: string
          description: The script language
        code:
          type: string
          description: The script code which has line endings properly escaped for JSON
    runScriptResponse:
      type: object
      properties:
        exitCode:
          type: integer
          description: The script process exit code - zero is no error
        stdout:
          type: string
          description: The stdout from the script process - scripts output data this way  
        stderr:
          type: string
          description: The stderr from the script process - scripts reports errors this way
    createMemoryRequest:
      type: object
      properties:
        id:
          type: string
          description: The id of the memory
        text:
          type: string
          description: The text of the memory
    createMemoryResponse:
      type: object
      properties:
        id:
          type: string
          description: The id of the created memory
    listMemoriesResponse:
      type: array
      items:
        type: string
      description: An array of memory ids
    getMemoryResponse:
      type: object
      properties:
        id:
          type: string
          description: The id of the memory
        text:
          type: string
          description: The text of the memory